finally decouple Config from the core. baahhahahahah
