+++ src/CMakeLists.txt
@@ -11,7 +11,6 @@ add_library(core STATIC
 	ARMInterpreter_ALU.cpp
 	ARMInterpreter_Branch.cpp
 	ARMInterpreter_LoadStore.cpp
-	Config.cpp
 	CP15.cpp
 	CRC32.cpp
 	DMA.cpp
