+++ src/frontend/Util_ROM.cpp
@@ -25,7 +25,6 @@
 #include "ArchiveUtil.h"
 #endif
 #include "FrontendUtil.h"
-#include "Config.h"
 #include "SharedConfig.h"
 #include "Platform.h"
