+++ src/frontend/qt_sdl/CheatsDialog.cpp
@@ -24,7 +24,6 @@
 #include "types.h"
 #include "Platform.h"
 #include "Config.h"
-#include "PlatformConfig.h"
 
 #include "CheatsDialog.h"
 #include "ui_CheatsDialog.h"
