+++ src/frontend/qt_sdl/Config.h
@@ -19,8 +19,6 @@
 #ifndef PLATFORMCONFIG_H
 #define PLATFORMCONFIG_H
 
-#include "Config.h"
-
 enum
 {
     HK_Lid = 0,
@@ -39,6 +37,17 @@ enum
 
 namespace Config
 {
+
+struct ConfigEntry
+{
+    char Name[32];
+    int Type;
+    void* Value;
+    int DefaultInt;
+    const char* DefaultStr;
+    int StrLength; // should be set to actual array length minus one
+};
+
 
 extern int KeyMapping[12];
 extern int JoyMapping[12];
@@ -128,6 +137,7 @@ extern int DirectLAN;
 extern int SavestateRelocSRAM;
 
 extern int AudioInterp;
+extern int AudioBitrate;
 extern int AudioVolume;
 extern int MicInputType;
 extern char MicWavPath[1024];
@@ -141,6 +151,10 @@ extern int EnableCheats;
 extern int MouseHide;
 extern int MouseHideSeconds;
 extern int PauseLostFocus;
+
+
+void Load();
+void Save();
 
 }
