+++ src/frontend/qt_sdl/EmuSettingsDialog.cpp
@@ -25,7 +25,6 @@
 #include "types.h"
 #include "Platform.h"
 #include "Config.h"
-#include "PlatformConfig.h"
 
 #include "EmuSettingsDialog.h"
 #include "ui_EmuSettingsDialog.h"
