+++ src/frontend/qt_sdl/FirmwareSettingsDialog.cpp
@@ -16,7 +16,7 @@
     with melonDS. If not, see http://www.gnu.org/licenses/.
 */
 
-#include "PlatformConfig.h"
+#include "Config.h"
 #include "FirmwareSettingsDialog.h"
 #include "ui_FirmwareSettingsDialog.h"
