+++ src/frontend/qt_sdl/InputConfig/InputConfigDialog.cpp
@@ -25,7 +25,6 @@
 
 #include "types.h"
 #include "Config.h"
-#include "PlatformConfig.h"
 
 #include "MapButton.h"
 #include "Input.h"
