+++ src/frontend/qt_sdl/WifiSettingsDialog.cpp
@@ -22,7 +22,6 @@
 #include "types.h"
 #include "Platform.h"
 #include "Config.h"
-#include "PlatformConfig.h"
 
 #include "LAN_Socket.h"
 #include "LAN_PCap.h"
